Flag to clean only succeeded clusters

+++ prow/scripts/lib/gardener/azure.sh
@@ -40,16 +40,18 @@ gardener::cleanup() {
     kubectl top pods --all-namespaces
 
     if [ -n "${CLEANUP_CLUSTER}" ]; then
-        log::info "Deprovision cluster: \"${CLUSTER_NAME}\""
-        # Export envvars for the script
-        export GARDENER_CLUSTER_NAME=${CLUSTER_NAME}
-        export GARDENER_PROJECT_NAME=${GARDENER_KYMA_PROW_PROJECT_NAME}
-        export GARDENER_CREDENTIALS=${GARDENER_KYMA_PROW_KUBECONFIG}
-        "${TEST_INFRA_CLUSTER_INTEGRATION_SCRIPTS}"/deprovision-gardener-cluster.sh
+        if  [ -z "${CLEANUP_ONLY_SUCCEEDED}" ] || [[ -n "${CLEANUP_ONLY_SUCCEEDED}" && ${EXIT_STATUS} -eq 0 ]]; then
+            log::info "Deprovision cluster: \"${CLUSTER_NAME}\""
+            # Export envvars for the script
+            export GARDENER_CLUSTER_NAME=${CLUSTER_NAME}
+            export GARDENER_PROJECT_NAME=${GARDENER_KYMA_PROW_PROJECT_NAME}
+            export GARDENER_CREDENTIALS=${GARDENER_KYMA_PROW_KUBECONFIG}
+            "${TEST_INFRA_CLUSTER_INTEGRATION_SCRIPTS}"/deprovision-gardener-cluster.sh
 
-        log::info "Deleting Azure EventHubs Namespace: \"${EVENTHUB_NAMESPACE_NAME}\""
-        # Delete the Azure Event Hubs namespace which was created
-        az eventhubs namespace delete -n "${EVENTHUB_NAMESPACE_NAME}" -g "${RS_GROUP}"
+            log::info "Deleting Azure EventHubs Namespace: \"${EVENTHUB_NAMESPACE_NAME}\""
+            # Delete the Azure Event Hubs namespace which was created
+            az eventhubs namespace delete -n "${EVENTHUB_NAMESPACE_NAME}" -g "${RS_GROUP}"
+        fi
     fi
 
     MSG=""
